+++ src/components/Modals/LoginToSaveModal.js
@@ -30,9 +30,7 @@ const LoginToSaveModal = () => {
       >
         <div className='modal-content__header'>
           <h2 className='modal-content__heading'>{t('loginToSaveModal.heading')}</h2>
-          <button onClick={closeModal}>
-            <CloseIcon/>
-          </button>
+          <Button className='btn--tertiary' onClickHandler={closeModal} ButtonIcon = {CloseIcon} />
         </div>
 
         <p className='modal-content__text'>{t('loginToSaveModal.loginText')}</p>