+++ src/components/ProjectIndex/ProjectIndex.js
@@ -10,6 +10,10 @@ import Button from '../Button/Button'
 import { createOrUpdateProject } from '../../utils/apiCallHandler'
 import { defaultPythonProject } from '../../utils/defaultProjects'
 import { PlusIcon } from '../../Icons';
+import RenameProjectModal from '../Modals/RenameProjectModal';
+import { showRenamedMessage } from '../../utils/Notifications';
+import { useEffect } from 'react';
+import DeleteProjectModal from '../Modals/DeleteProjectModal';
 
 const ProjectIndex = (props) => {
   const history = useHistory();
@@ -19,6 +23,15 @@ const ProjectIndex = (props) => {
   useRequiresUser(isLoading, user);
   useProjectList(user);
   const projectListLoaded = useSelector((state) => state.editor.projectListLoaded);
+  const renameProjectModalShowing = useSelector((state) => state.editor.renameProjectModalShowing)
+  const deleteProjectModalShowing = useSelector((state) => state.editor.deleteProjectModalShowing)
+  const saving = useSelector((state) => state.editor.saving)
+
+  useEffect(() => {
+    if (saving === 'success') {
+      showRenamedMessage()
+    }
+  }, [saving])
 
   const onCreateProject = async () => {
     const response = await createOrUpdateProject(defaultPythonProject, user.access_token);
@@ -38,7 +51,11 @@ const ProjectIndex = (props) => {
           ButtonIcon={PlusIcon}
         />
       </ProjectIndexHeader>
-      { projectListLoaded === true ? <ProjectListTable /> : <p>{t('projectList.loading')}...</p> }
+      { projectListLoaded === 'success' ? <ProjectListTable /> :
+        projectListLoaded === 'failed' ? <p>{t('projectList.loadingFailed')}</p> :
+        <p>{t('projectList.loading')}</p> }
+      { renameProjectModalShowing ? <RenameProjectModal /> : null }
+      { deleteProjectModalShowing ? <DeleteProjectModal /> : null }
     </>
   );
 };